python: str.decode() doesn't like kwargs in python 2.5
Recent changes introduced lots of unicodification of strings, mostly in
the form of .decode('utf-8', errors='ignore'). However, python 2.5 does
not like the errors keyword argument and complains. It does work when
used as a simple arg though, so that's what this patch does.

Signed-off-by: Sebastian Spaeth <Sebastian@SSpaeth.de>

Backported to 0.11 by David Bremner <david@tethera.net>

Conflicts:

	bindings/python/notmuch/database.py
	bindings/python/notmuch/filename.py
	bindings/python/notmuch/message.py
	bindings/python/notmuch/thread.py