build system: remove configure output in Make distclean.
Create a variable DISTCLEAN which contains a list of things to
clean in the distclean target (in addition to running the clean
target).

The deleted comment seems to be false these days, since we do
create files during configuration.

Use "rm -rf" here as well in case we want to add directories to
DISTCLEAN.